More Reorganization Plus Some Recolors
Reorganized many elements in the main xcf and in the folders to be
sorted better and easier to find. Also created a new pair of bracers (I
missed the ones already there, which is also why I reorganized), and
made recolors for both pants and shoes. I also renamed the existing
pants colors from green and red to teal and magenta (the closest colors
I could think of).