Added Female Sleeveless Shirt Sprites
Duplicated the female pirate shirt sprites and modified it to be a
full-fledged sleeveless shirt (covering the entire neck area instead of
just being spaghetti strings) and also added a maroon recolor of the
shirt. Also moved Tabard outside of Chainmail and moved the Tabard
colors to the root Shirts folder in the XCF. Next step: capes.