Added Neckpieces for Capes in Accessories
First of the cape sprites, the neckpiece. The capes need not necessarily
have neck pieces, they may still work without them, but the neckpiece is
to show where the cape is being held up.

There are two variants so far; tied and clipped. There are a number of
recolors for them, although I'm not sure I like the looks of every
recolor. Additional neckpieces are planned later on, but this is what
I've got so far.