Added Back Cape Sprites
Added cape sprites for the back, both behind body and in front. The
front sprites are currently located in torso/back/cape. There are a
number of recolors for them, as well as a trimmed version of the cape
(the trim only appears on the outside of the cape).

I like this for the most part, though I'm not sure I like all the
animations. I think the side animations may need some touching up, and
perhaps the north animations as well. I love how the base sprites turned
out though. I'm not confident in all the recolors, either, but I like
most of them.

More cape variants will be out later, but for now I only have trimmed
(only one color), and normal. Tattered is the next one I'm working on.