Fixed Hurt Frames for Loose Hairstyle
Made adjustments to three frames on the loose hairstyle to fix it up.
These are the third, fourth, and fifth frames of the hurt animation. The
third animation is merely extended to be slightly taller, the fourth has
more shading at the top, and the fifth has been heavily updated to
properly depict the top of the hair.

I also added a new hair color light-blonde2, which is for my original
character Maya Leudona. The existing hair colors weren't quite close
enough to her shade of blonde for my liking, so I created a new one with
the palette from an existing artwork I had of her.