Finished Red Orc Recolors
These don't look much like "red" orc body sprites, but whatever the case
may be I did a full recolor of the base orc sprites to the red orc
sprites for both male and female. The female sprites are the main one,
but I added male sprites just to be consistent.